backend/requirements.txt
aniso8601==6.0.0
Click==7.0
Flask==1.0.3
Flask-Cors==3.0.7
Flask-RESTful==0.3.7
Flask-SQLAlchemy==2.4.0
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
# I changed from 2.8.2 to 2.8.6 because I
# kept getting an error during pip install
# that said it could find a version 2.8.2.
psycopg2-binary==2.8.6
pytz==2019.1
six==1.12.0
SQLAlchemy==1.3.4
# Based on StackOverlow answer, I changed from
# 0.15.4 to 0.15.5 due to error I was getting.
# https://stackoverflow.com/a/60140477
Werkzeug==0.15.5
python-dotenv==0.15.0
